- Remove the temporary gRPC log channel - Add consideration for log filterers - Modify all affected Docs
