+++ README.md
@@ -56,7 +56,7 @@ Alternatively, you can also pass the configurations via environment variables (s
 All supported environment variables can be found [here](docs/EnvVars.md)
 
 ## Report logs with Python Agent
-The Python agent is capable of reporting collected logs to the backend(SkyWalking OAP/ [SkyWalking Satellite Sidecar](https://github.com/apache/skywalking-satellite)), enabling Log & Trace Correlation.
+The Python agent is capable of reporting collected logs to the backend(SkyWalking OAP), enabling Log & Trace Correlation.
 
 Please refer to the [Log Reporter Doc](docs/LogReporter.md) for a detailed guide.
