+++ skywalking/agent/__init__.py
@@ -26,7 +26,6 @@ from skywalking.protocol.logging.Logging_pb2 import LogData
 from skywalking import config, plugins
 from skywalking import loggings
 from skywalking.agent.protocol import Protocol
-from skywalking.agent.protocol.grpc_log import GrpcLogProtocol
 from skywalking.command import command_service
 from skywalking.config import profile_active, profile_task_query_interval
 from skywalking.loggings import logger
@@ -35,7 +34,7 @@ if TYPE_CHECKING:
     from skywalking.trace.context import Segment
 
 __started = False
-__protocol = __log_protocol = Protocol()  # type: Protocol
+__protocol = Protocol()  # type: Protocol
 __heartbeat_thread = __report_thread = __log_report_thread = __query_profile_thread = __command_dispatch_thread \
     = __queue = __log_queue = __finished = None
 
@@ -60,20 +59,10 @@ def __report():
         __finished.wait(0)
 
 
-def __log_heartbeat():
+def __report_log():
     while not __finished.is_set():
         try:
-            __log_protocol.heartbeat()
-        except Exception as exc:
-            logger.error(str(exc))
-
-        __finished.wait(30)
-
-
-def __log_report():
-    while not __finished.is_set():
-        try:
-            __log_protocol.report(__log_queue)
+            __protocol.report_log(__log_queue)
         except Exception as exc:
             logger.error(str(exc))
 
@@ -112,9 +101,7 @@ def __init_threading():
 
     if config.log_grpc_reporter_active:
         __log_queue = Queue(maxsize=config.log_grpc_reporter_max_buffer_size)
-        __log_heartbeat_thread = Thread(name='LogHeartbeatThread', target=__log_heartbeat, daemon=True)
-        __log_report_thread = Thread(name='LogReportThread', target=__log_report, daemon=True)
-        __log_heartbeat_thread.start()
+        __log_report_thread = Thread(name='LogReportThread', target=__report_log, daemon=True)
         __log_report_thread.start()
 
     if profile_active:
@@ -122,7 +109,7 @@ def __init_threading():
 
 
 def __init():
-    global __protocol, __log_protocol
+    global __protocol
 
     if config.protocol == 'grpc':
         from skywalking.agent.protocol.grpc import GrpcProtocol
@@ -135,9 +122,8 @@ def __init():
         __protocol = KafkaProtocol()
 
     plugins.install()
-    if config.log_grpc_reporter_active:
+    if config.log_grpc_reporter_active:  # todo - Add support for printing traceID/ context in logs
         from skywalking import log
-        __log_protocol = GrpcLogProtocol()
         log.install()
 
     __init_threading()
@@ -147,7 +133,7 @@ def __fini():
     __protocol.report(__queue, False)
     __queue.join()
     if config.log_grpc_reporter_active:
-        __log_protocol.report(__log_queue, False)
+        __protocol.report_log(__log_queue, False)
         __log_queue.join()
     __finished.set()
